requirements.txt
# Core engine
typer
pydantic
requests
pyyaml
google-genai
python-dotenv

# Telegram bot
aiogram>=3.3.0
aiosqlite>=0.19.0

# Dashboard
fastapi>=0.110.0
python-multipart>=0.0.6
uvicorn>=0.27.0
jinja2>=3.1.0
